+++ netbox_healthcheck_plugin/navigation.py
@@ -1,10 +1,5 @@
-from netbox.plugins import PluginMenuButton, PluginMenuItem
-
+from netbox.plugins import PluginMenuItem
 
 menu_items = (
-    PluginMenuItem(
-        link='plugins:netbox_healthcheck_plugin:healthcheck_list',
-        link_text='HealthCheck',
-        buttons=None
-    ),
+    PluginMenuItem(link='plugins:netbox_healthcheck_plugin:healthcheck_list', link_text='HealthCheck', buttons=None),
 )