+++ tests/test_netbox_healthcheck_plugin.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-
-"""Tests for `netbox_healthcheck_plugin` package."""
-
-from netbox_healthcheck_plugin import netbox_healthcheck_plugin